+++ lib/elasticquery.rb
@@ -3,4 +3,5 @@ require "elasticquery/version"
 # Base gem module
 module Elasticquery
   autoload :Base, "elasticquery/base"
+  autoload :Es, "elasticquery/es"
 end