+++ History.md
@@ -10,6 +10,8 @@
 
 * Add `without`/`missing` filter
 
+* Add `es` shortcut from `Elasticquery::Es` to model chain calls
+
 ## Elasticquery 0.1.2 (16 June 2015)
 
 * Fix `range.not`. Exception was raised